gui is functional for OpenCL
No validation yet, but you can render and save a fractal generated with
OCL

+++ libpbrot/PbrotCommon.h
@@ -12,6 +12,6 @@
 #define CLfraction_t float
 #define OMPfraction_t double
 
-uint16_t* normalizeCLGrid(CLbucket_t* grid, int gridSize);
+uint8_t* normalizeCLGrid(CLbucket_t* grid, int gridSize);
 
 void writeImage(int width, int height, uint16_t* buffer);
